refactor(day-diet): introduce day diet gateway pattern
